@@ -14,8 +14,7 @@
 	<name>ecommerce</name>
 	<description>Ecommerce project for Spring Boot</description>
 	<properties>
-		<java.version>1.8</java.version>
-		
+		<java.version>11</java.version>
 	</properties>
 	
 	<dependencies>
