@@ -2,8 +2,6 @@ package com.tsk.ecommerce.controllers;
 
 import javax.validation.Valid;
 
-import org.springframework.beans.factory.annotation.Autowired;
-
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,16 +13,17 @@ import com.tsk.ecommerce.services.messageries.sms.SmsSender;
 
 @RestController
 @RequestMapping("api/v1")
-public class SmsResourse {
+public class SmsController {
 
-	@Autowired
-	private SmsSender smsSender;
-	
+	private final SmsSender smsSender;
+
+	public SmsController(SmsSender smsSender) {
+		this.smsSender = smsSender;
+	}
+
 	@PostMapping("/sms")
 	public void sendSms(@Valid @RequestBody SmsRequest smsRequest) {
-		
 		smsSender.sendSms(smsRequest);
-		
 	}
 	
 }
