@@ -12,9 +12,15 @@ import javax.persistence.ManyToOne;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+@Data
 @Entity
 public class OrderLine implements Serializable {
 	
@@ -30,61 +36,5 @@ public class OrderLine implements Serializable {
 	@ManyToOne
 	@JoinColumn(name = "id_order")
 	private Orders order;
-
-	public OrderLine() {
-		super();
-	}
-
-
-	public OrderLine(Integer quantity, Product product) {
-		super();
-		this.quantity = quantity;
-		this.product = product;
-	}
-
-
-	public Double getSubTotal() {
-		return subTotal;
-	}
-
-
-	public void setSubTotal(Double subTotal) {
-		this.subTotal = subTotal;
-	}
-
-
-	public Long getIdOrderLine() {
-		return idOrderLine;
-	}
-
-	public void setIdOrderLine(Long idOrderLine) {
-		this.idOrderLine = idOrderLine;
-	}
-
-	public Integer getQuantity() {
-		return quantity;
-	}
-
-	public void setQuantity(Integer quantity) {
-		this.quantity = quantity;
-	}
-
-
-	public Product getProduct() {
-		return product;
-	}
-
-	public void setProduct(Product product) {
-		this.product = product;
-	}
-
-
-	@Override
-	public String toString() {
-		return "OrderLine [idOrderLine=" + idOrderLine + ", quantity=" + quantity + ", total=" + subTotal + ", product="
-				+ product + "]";
-	}
-	
-	
 	
 }
