@@ -4,11 +4,11 @@ import com.tsk.ecommerce.dtos.requests.SignUpRequest;
 import com.tsk.ecommerce.entities.auth.RoleEntity;
 import com.tsk.ecommerce.entities.auth.UserEntity;
 import com.tsk.ecommerce.entities.enumerations.ERole;
-import com.tsk.ecommerce.exceptions.FormatDataInvalidException;
 import com.tsk.ecommerce.exceptions.ResourceNotFoundException;
 import com.tsk.ecommerce.repositories.RoleEntityRepository;
 import com.tsk.ecommerce.repositories.UserEntityRepository;
-import com.tsk.ecommerce.services.validations.UserValidation;
+import com.tsk.ecommerce.validators.UserValidator;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -20,30 +20,24 @@ public class AccountService {
 
     private final PasswordEncoder passWordEncoder;
     private final UserEntityRepository userEntityRepository;
-    private final UserValidation userValidation;
+    private final UserValidator userValidator;
     private final RoleEntityRepository roleEntityRepository;
 
     public AccountService(PasswordEncoder passWordEncoder,
                           UserEntityRepository userEntityRepository,
-                          UserValidation userValidation,
+                          UserValidator userValidator,
                           RoleEntityRepository roleEntityRepository) {
         this.passWordEncoder = passWordEncoder;
         this.userEntityRepository = userEntityRepository;
-        this.userValidation = userValidation;
+        this.userValidator = userValidator;
         this.roleEntityRepository = roleEntityRepository;
     }
 
-    public UserEntity register(SignUpRequest request) throws IOException {
+    public UserEntity registerPhase1(SignUpRequest request) throws IOException {
+        userValidator.validateSignUp(request);
         UserEntity newUser = new UserEntity();
-        if (!userValidation.isUsernameExisted(request.getUsername())) {
-            newUser.setUsername(request.getUsername());
-        } else throw new FormatDataInvalidException("User already used");
-
-        if (userValidation.isEmailExisted(request.getEmail())) throw new FormatDataInvalidException("Email already used");
-        else {
-            newUser.setEmail(request.getEmail());
-        }
-
+        newUser.setUsername(request.getUsername());
+        newUser.setEmail(request.getEmail());
         RoleEntity role = roleEntityRepository.findByName(ERole.ROLE_USER).get();
         newUser.setRoles(Collections.singletonList(role));
         newUser.setPassword(passWordEncoder.encode(request.getPassword()));
@@ -52,9 +46,8 @@ public class AccountService {
 
     public UserEntity getByUsernameAndPassword(String username, String password) {
         UserEntity user = this.getByUsername(username);
-        if(passWordEncoder.matches(password, user.getPassword())) {
-            return user;
-        }else throw new ResourceNotFoundException("Echec d'authentification !");
+        if(!passWordEncoder.matches(password, user.getPassword())) throw new ResourceNotFoundException("Echec d'authentification !");
+        return user;
     }
 
     public UserEntity getByUsername(String username) {
