@@ -4,7 +4,6 @@ import com.tsk.ecommerce.entities.Product;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import java.util.List;
@@ -14,17 +13,19 @@ import static org.junit.jupiter.api.Assertions.*;
 @DataJpaTest
 class ProductRepositoryTest {
 
-    @Autowired
-    private ProductRepository productRepository;
+    private final ProductRepository productRepository;
+
+    ProductRepositoryTest(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
 
     @BeforeEach
     void setUp() {
-        Product inputProduct = Product.builder()
-                .nameProduct("KeyBoard")
-                .description("AZERTY alignment")
-                .price(400.0)
-                .stock(10)
-                .build();
+        Product inputProduct = new Product();
+        inputProduct.setNameProduct("KeyBoard");
+        inputProduct.setDescription("AZERTY alignment");
+        inputProduct.setPrice(400.0);
+        inputProduct.setStock(10);
         productRepository.save(inputProduct);
     }
 
